+++ PE1/LinearFunction.h
@@ -4,14 +4,14 @@
 class LinearFunction : public FunctionContainer
 {
 private:
-	float _k;
-	float _b;
+	double _k;
+	double _b;
 
 public:
-	LinearFunction(float k, float b) {
+	LinearFunction(double k, double b) {
 		_k = k;
 		_b = b;
 	}
-	float ComputeWith(float x) override;
+	double ComputeWith(double x) override;
 	std::string ToString() override;
 };
